+++ frontend/src/metabase-lib/lib/Alert/index.ts
@@ -1,0 +1,7 @@
+export {
+  ALERT_TYPE_ROWS,
+  ALERT_TYPE_TIMESERIES_GOAL,
+  ALERT_TYPE_PROGRESS_BAR_GOAL,
+} from "./constants";
+
+export { getDefaultAlert } from "./Alert";
